extract statusBar as element
